hospital-system/doctor-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
email-validator==2.2.0
prometheus-fastapi-instrumentator==7.0.0
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
